refactor(agent_backends): extract auto_detect submodule
Move startup auto-detect / probation / opt-out into auto_detect.rs:

- Constants: AUTO_DETECT_DISABLED_PREFIX, AUTO_DETECT_TIMEOUT,
  PI_AUTO_DETECT_TIMEOUT (pi-sdk only)
- BackendAutoDetection struct
- auto_detect_disabled_key, backend_supports_auto_detect,
  backend_auto_detect_disabled, should_probe_backend_auto_detection
- skipped_backend_auto_detection,
  persist_backend_auto_detect_opt_out,
  apply_backend_auto_detections
- probe_codex_backend, probe_model_discovery_backend,
  backend_auto_detect_timeout, codex_startup_models

The probe helpers depend on discover_models / discover_codex_models /
codex_cli_command which still live in the facade. Those three are
exposed as pub(super) in mod.rs for now; commit 6 (discovery
extraction) will move them into discovery.rs alongside the rest of
the per-backend discovery cluster, at which point the pub(super)
markers naturally relocate.

Tests covering auto_detect helpers still live in mod.rs's `mod tests`
for this commit and reach the pub(super) functions via `use
super::auto_detect::*` — a focused test-colocation pass follows when
the facade is trimmed. PI_AUTO_DETECT_TIMEOUT is gated behind
`#[cfg(feature = "pi-sdk")]` in the test imports to match the
production gate.

cargo fmt --check, cargo clippy -p claudette -p claudette-server -p
claudette-cli --all-targets --all-features with -Dwarnings, and
the agent_backends tests (111 passed) all pass. Test count matches
baseline; no public-surface change.
